refactor(notes): separa api/ en actions/ y repository/
