+++ src/features/notes/api/repository/libraryOfArchivedNotesNextjsRepository.ts
@@ -1,19 +1,19 @@
 import {
 	defaultLibraryOfArchivedNotes,
 	LibraryOfArchivedNotes,
-} from '../model/libraryOfArchivedNotes'
+} from '../../model/libraryOfArchivedNotes'
 import { LibraryOfArchiveNotesRepository } from './libraryOfArchivedNotesRepository'
 
 export default class LibraryOfArchivedNotesNextjsRepository
 	implements LibraryOfArchiveNotesRepository
 {
 	async save(library: LibraryOfArchivedNotes, boardId: string): Promise<void> {
-		const { saveArchivedNotes } = await import('./saveArchivedNotes')
+		const { saveArchivedNotes } = await import('../actions/saveArchivedNotes')
 		await saveArchivedNotes({ boardId, notes: library })
 	}
 
 	async getAll(boardId: string): Promise<LibraryOfArchivedNotes> {
-		const { getArchivedNotes } = await import('./getArchivedNotes')
+		const { getArchivedNotes } = await import('../actions/getArchivedNotes')
 		const notes = await getArchivedNotes({ boardId })
 		return notes ?? defaultLibraryOfArchivedNotes
 	}
