+++ src/features/notes/api/repository/libraryOfArchivedNotesRepositoryFactory.ts
@@ -1,5 +1,5 @@
 import type { SessionType } from '@/auth/contexts/SessionProvider'
-import { LibraryOfArchivedNotes } from '../model/libraryOfArchivedNotes'
+import { LibraryOfArchivedNotes } from '../../model/libraryOfArchivedNotes'
 import LibraryOfArchivedNotesLocalStorageRepository from './libraryOfArchivedNotesLocalStorageRepository'
 import LibraryOfArchivedNotesNextjsRepository from './libraryOfArchivedNotesNextjsRepository'
 import { LibraryOfArchiveNotesRepository } from './libraryOfArchivedNotesRepository'
