+++ src/features/notes/api/repository/nextjsNotesRepository.ts
@@ -1,14 +1,14 @@
-import { defaultNotes, Notes } from '../model/notes'
+import { defaultNotes, Notes } from '../../model/notes'
 import { NotesRepository } from './notesRepository'
 
 export default class NextjsNotesRepository implements NotesRepository {
 	async save(notes: Notes, boardId: string): Promise<void> {
-		const { saveNotes } = await import('./saveNotes')
+		const { saveNotes } = await import('../actions/saveNotes')
 		await saveNotes({ boardId, notes })
 	}
 
 	async getAll(boardId: string): Promise<Notes> {
-		const { getNotes } = await import('./getNotes')
+		const { getNotes } = await import('../actions/getNotes')
 		const notes = await getNotes({ boardId })
 		return notes ?? defaultNotes
 	}
